fix: Move ln-gateway/src/gatewayd contents to parent folder
