+++ gateway/ln-gateway/src/gateway.rs
@@ -22,8 +22,8 @@ use tracing::{error, info, warn};
 
 use super::actor::GatewayActor;
 use crate::client::DynGatewayClientBuilder;
-use crate::gatewayd::lnrpc_client::DynLnRpcClient;
 use crate::gatewaylnrpc::GetPubKeyResponse;
+use crate::lnrpc_client::DynLnRpcClient;
 use crate::rpc::rpc_server::run_webserver;
 use crate::rpc::{
     BackupPayload, BalancePayload, ConnectFedPayload, DepositAddressPayload, DepositPayload,
