+++ gateway/tests/tests/fixtures/mod.rs
@@ -8,8 +8,8 @@ use fedimint_testing::btc::fixtures::FakeBitcoinTest;
 use fedimint_testing::btc::BitcoinTest;
 use fedimint_testing::ln::fixtures::FakeLightningTest;
 use ln_gateway::client::{DynGatewayClientBuilder, MemDbFactory};
-use ln_gateway::gatewayd::gateway::Gateway;
-use ln_gateway::gatewayd::lnrpc_client::DynLnRpcClient;
+use ln_gateway::gateway::Gateway;
+use ln_gateway::lnrpc_client::DynLnRpcClient;
 use mint_client::module_decode_stubs;
 use mint_client::modules::wallet::WalletGen;
 use url::Url;
