+++ integrationtests/tests/fixtures/mod.rs
@@ -52,10 +52,10 @@ use futures::future::{join_all, select_all};
 use futures::{FutureExt, StreamExt};
 use hbbft::honey_badger::Batch;
 use itertools::Itertools;
+use ln_gateway::actor::GatewayActor;
 use ln_gateway::client::{DynGatewayClientBuilder, MemDbFactory, StandardGatewayClientBuilder};
-use ln_gateway::gatewayd::actor::GatewayActor;
-use ln_gateway::gatewayd::gateway::Gateway;
-use ln_gateway::gatewayd::lnrpc_client::{DynLnRpcClient, NetworkLnRpcClient};
+use ln_gateway::gateway::Gateway;
+use ln_gateway::lnrpc_client::{DynLnRpcClient, NetworkLnRpcClient};
 use mint_client::mint::SpendableNote;
 use mint_client::transaction::legacy::Transaction;
 use mint_client::transaction::TransactionBuilder;
